Lexer can now scan for A,E,F,G,U,X.
For now these are hooks for later use.  In particular,
the parser cannot use those tokens yet.  We will first
need to define syntax for CTL* and other logics, and
decide where we want to allow them to appear.  Also,
there is an option "ParseTemporalOperators", defaults
to "false", which turns off scanning for these tokens.